+++ api/package.json
@@ -55,11 +55,11 @@
   },
   "devDependencies": {
     "@loopback/build": "^6.2.0",
-    "@loopback/eslint-config": "^9.0.0",
+    "source-map-support": "^0.5.19",
     "@loopback/testlab": "^3.2.2",
     "@types/node": "^10.17.28",
+    "@loopback/eslint-config": "^9.0.0",
     "eslint": "^7.6.0",
-    "source-map-support": "^0.5.19",
     "typescript": "~3.9.7"
   }
 }
