+++ client/src/api/models/Route.ts
@@ -1,0 +1,6 @@
+import {Station} from './Station';
+
+export interface Route {
+	cost: number;
+	stations: Station[];
+}
