+++ client/src/components/RouteDisplay.tsx
@@ -1,9 +1,11 @@
-import React, { FunctionComponent} from 'react';
+import React, {FunctionComponent} from 'react';
 
 interface RouteDisplayProps {
-
+	test?: string;
 }
 
 export const RouteDisplay: FunctionComponent<RouteDisplayProps> = (props: RouteDisplayProps): JSX.Element => {
-	return ()
-}
+	return (
+		<div/>
+	);
+};
