+++ client/src/components/StationSelect.tsx
@@ -2,7 +2,7 @@ import {Select} from 'antd';
 import React from 'react';
 import {OptionData, OptionGroupData, OptionsType} from 'rc-select/lib/interface';
 
-import {Station} from '../models/Station';
+import {Station} from '../api/models/Station';
 
 interface StationAutoCompleteProps {
 	onChange: (id:number) => void;
